Centralize secrets detection test helpers
