+++ plugins/tests/secrets_detection/helpers.py
@@ -1,0 +1,69 @@
+from pydantic import BaseModel, RootModel, model_serializer
+
+from mcpgateway.common.models import ResourceContent
+from mcpgateway.plugins.framework import (
+    PluginConfig,
+    PluginContext,
+    PluginManager,
+    PluginMode,
+    PromptHookType,
+    PromptPrehookPayload,
+    ResourceHookType,
+    ResourcePostFetchPayload,
+    ToolHookType,
+    ToolPostInvokePayload,
+)
+from mcpgateway.plugins.framework.models import GlobalContext
+
+from cpex_secrets_detection.secrets_detection import SecretsDetectionPlugin
+from cpex_secrets_detection.secrets_detection_rust import py_scan_container
+
+__all__ = [
+    "BaseModel",
+    "GlobalContext",
+    "PluginManager",
+    "PluginMode",
+    "PromptHookType",
+    "PromptPrehookPayload",
+    "ResourceContent",
+    "ResourceHookType",
+    "ResourcePostFetchPayload",
+    "RootModel",
+    "SecretsDetectionPlugin",
+    "ToolHookType",
+    "ToolPostInvokePayload",
+    "_make_config",
+    "_make_context",
+    "make_config",
+    "make_context",
+    "model_serializer",
+    "py_scan_container",
+]
+
+
+def make_context() -> PluginContext:
+    return PluginContext(
+        global_context=GlobalContext(request_id="req-secrets", server_id="srv-secrets")
+    )
+
+
+def make_config(**overrides) -> PluginConfig:
+    config = {
+        "block_on_detection": False,
+        "redact": True,
+        "redaction_text": "[REDACTED]",
+    }
+    config.update(overrides)
+    return PluginConfig(
+        name="secrets_detection",
+        kind="cpex_secrets_detection.secrets_detection.SecretsDetectionPlugin",
+        config=config,
+    )
+
+
+def _make_context() -> PluginContext:
+    return make_context()
+
+
+def _make_config(**overrides) -> PluginConfig:
+    return make_config(**overrides)
